Add DTO & Refactoring
Need to remove logic from controller because it should not be there, that logic should be in the Service Layer or somewhere.

Need to implement DTO better in other Services and refactor POST and PUT to use DTO when needed.

JSON views removed because DTO is in use.

Still need more refactoring and testing.